Added ability for officers to create applications on behalf of a user.
New step between selecting licence type and selecting profile, which is
to select or create a customer.

New proxy applicant attribute on applications.

Using a common header for licence entry pages that will display the
customer name, if available.

Refactor of view classes in entry to use get_context_data, rather than
get, along with a base class that returns the customer and licence type
as context data.