Various Bugfixes
Fixed but with cover letter not appearing in PDF.

ID Updates / Amendments / Return Requests now email to proxy user if the
application is a on-behalf application.

Officers can now renew licences on behalf.

Fixed Is Renewable checkbox being different for issue/reissue.

Fixed but with saving profiles in customer management.