WIP Numerous bugfixes
#1837 Review of Returns not required for amendment appl

Officers will no longer see return status of previous application nor
have the returns item on the processing checklist for licence
amendments.

#1846 Confirmation once ID has been uploaded

This has now been fixed so that if a user uploads ID and there is an
application awaiting ID upload, the user will see a message saying that
the application processing will proceed.