Added renewal_sent flag to renewals
Flag is set to true when either an email has been sent for a renewal
reminder or a renewal pdf has been created for a licence.

Renewals aren't sent for licences that have been amended / renewed
already.

A re-issue of a licence will set the flag to back to false allowing
renewal reminders to be sent for reissued licences.

Changed licences 'previous_licence' flag, which wasn't being used
(except for being shown in reports, which would always be empty) to
'replaced_by', which is used as part of the condition as to whether to
send renewal reminders.

Moved renewal reminder cron job and email code out of applications app
into main, where they are more relevant.