Added ability to save an application during the processing steps in the enter conditions / issue licence pages.
Licences are now created for applications in the issue licence page
whether you save or issue.

The difference is that the issue steps and the status changes only get
applied when you issue - if you save this just saves the fields
associated with a licence, such as additional info.

Updated other parts of the site, such as customer dashboard, to reflect
the changes.

Added unit tests.